CI: Various OSX build fixes

+++ tools/travis/travis_osx_script.sh
@@ -19,6 +19,7 @@ mkdir -p build
 cd build
 cmake -DCMAKE_BUILD_TYPE=Debug -DUA_NAMESPACE_ZERO=FULL -DUA_BUILD_EXAMPLES=ON  ..
 make -j
+if [ $? -ne 0 ] ; then exit 1 ; fi
 cd .. && rm -rf build
 echo -en 'travis_fold:end:script.build.ns0\\r'
 
@@ -26,7 +27,9 @@ echo "Compile release build for OS X" && echo -en 'travis_fold:start:script.buil
 mkdir -p build && cd build
 cmake -DCMAKE_BUILD_TYPE=Release -DUA_ENABLE_AMALGAMATION=OFF -DUA_BUILD_EXAMPLES=ON -DCMAKE_INSTALL_PREFIX=${TRAVIS_BUILD_DIR}/open62541-osx ..
 make -j
+if [ $? -ne 0 ] ; then exit 1 ; fi
 make install
+if [ $? -ne 0 ] ; then exit 1 ; fi
 cd ..
 tar -pczf open62541-osx.tar.gz LICENSE AUTHORS README.md ${TRAVIS_BUILD_DIR}/open62541-osx/*
 rm -rf build
@@ -36,12 +39,14 @@ echo "Compile multithreaded version" && echo -en 'travis_fold:start:script.build
 mkdir -p build && cd build
 cmake -DUA_ENABLE_MULTITHREADING=ON -DUA_BUILD_EXAMPLES=ON ..
 make -j
+if [ $? -ne 0 ] ; then exit 1 ; fi
 cd .. && rm -rf build
 echo -en 'travis_fold:end:script.build.multithread\\r'
 
 echo "Debug build and unit tests with valgrind" && echo -en 'travis_fold:start:script.build.unit_test\\r'
 mkdir -p build && cd build
-cmake -DCHECK_PREFIX=/usr/local/Cellar/check/0.11.0 -DCMAKE_BUILD_TYPE=Debug -DUA_BUILD_EXAMPLES=ON -DUA_ENABLE_ENCRYPTION=ON -DUA_ENABLE_DISCOVERY=ON -DUA_ENABLE_DISCOVERY_MULTICAST=ON -DUA_BUILD_UNIT_TESTS=ON -DUA_ENABLE_COVERAGE=ON -DUA_ENABLE_UNIT_TESTS_MEMCHECK=ON ..
+cmake -DCHECK_PREFIX=/usr/local/Cellar/check/0.11.0 -DCMAKE_BUILD_TYPE=Debug -DUA_BUILD_EXAMPLES=ON -DUA_ENABLE_ENCRYPTION=ON -DUA_ENABLE_DISCOVERY=ON -DUA_ENABLE_DISCOVERY_MULTICAST=ON -DUA_BUILD_UNIT_TESTS=ON -DUA_ENABLE_COVERAGE=OFF -DUA_ENABLE_UNIT_TESTS_MEMCHECK=ON ..
 make -j && make test ARGS="-V"
+if [ $? -ne 0 ] ; then exit 1 ; fi
 cd .. && rm -rf build
 echo -en 'travis_fold:end:script.build.unit_test\\r'
